+++ docx_generator/lib/src/xml_parts/xml_parts.dart
@@ -1,5 +1,6 @@
 export 'content_types_xml.dart';
 export 'document_xml.dart';
+export 'header_footer_xml.dart';
 export 'numbering_xml.dart';
 export 'rels_xml.dart';
 export 'styles_xml.dart';